Move `UnsupportedTypeError`
