+++ packages/seroval/src/UnsupportedTypeError.ts
