+++ packages/seroval/src/tree/sync.ts
@@ -5,7 +5,7 @@ import type { ParserContext } from './context';
 import { createIndexedValue } from './context';
 import { serializeString } from '../string';
 import type { BigIntTypedArrayValue, TypedArrayValue } from '../types';
-import UnsupportedTypeError from './UnsupportedTypeError';
+import UnsupportedTypeError from '../UnsupportedTypeError';
 import {
   TRUE_NODE,
   FALSE_NODE,
